+++ src/neptune/new/internal/utils/deprecation.py
@@ -22,7 +22,7 @@ from neptune.new.exceptions import NeptuneParametersCollision
 __all__ = ["deprecated", "deprecated_parameter"]
 
 
-def deprecated(*, alternative: Optional[str] = None, stack_level: int = 1):
+def deprecated(*, alternative: Optional[str] = None):
     def deco(func):
         @wraps(func)
         def inner(*args, **kwargs):
@@ -30,8 +30,8 @@ def deprecated(*, alternative: Optional[str] = None, stack_level: int = 1):
 
             warn_once(
                 message=f"`{func.__name__}` is deprecated{additional_info}."
-                f" We'll end support of it in `neptune-client==1.0.0`.",
-                stack_level=stack_level + 1,
+                f" We'll end support of it in `neptune-client==1.0.0`."
+                " For details, see https://docs.neptune.ai/setup/neptune-client_1-0_release_changes"
             )
 
             return func(*args, **kwargs)
@@ -51,8 +51,7 @@ def deprecated_parameter(*, deprecated_kwarg_name, required_kwarg_name):
 
                 warn_once(
                     message=f"Parameter `{deprecated_kwarg_name}` is deprecated, use `{required_kwarg_name}` instead."
-                    " We'll end support of it in `neptune-client==1.0.0`.",
-                    stack_level=2,
+                    " We'll end support of it in `neptune-client==1.0.0`."
                 )
 
                 kwargs[required_kwarg_name] = kwargs[deprecated_kwarg_name]