+++ CHANGELOG.md
@@ -1,6 +1,15 @@
-## neptune-client 0.16.16 [UNRELEASED]
+## neptune-client 0.16.16
+
+### Features
+- Added `stringify_unsupported` function for handling backward compatibility of implicit casting ([#1177](https://github.com/neptune-ai/neptune-client/pull/1177))
+- Better support for `Handler` level objects ([#1178](https://github.com/neptune-ai/neptune-client/pull/1178))
+
+### Changes
+- Docstrings and deprecation messages updated ([#1182](https://github.com/neptune-ai/neptune-client/pull/1182))
+- Deprecate name parameter in init_project and management API ([#1175](https://github.com/neptune-ai/neptune-client/pull/1175))
 
 ### Fixes
+- Fixed deprecation warnings for implicit casting to string ([#1177](https://github.com/neptune-ai/neptune-client/pull/1177))
 - Disabled info about stopping when using read-only mode ([#1166](https://github.com/neptune-ai/neptune-client/pull/1166))
 - Disabled "Explore the metadata" message when stopping in debug mode ([#1165](https://github.com/neptune-ai/neptune-client/pull/1165))